updated deleting user screen
